operator admin

+++ src/main/java/root/json/RateResponseEntity.java
@@ -16,7 +16,7 @@ public class RateResponseEntity
     @JsonProperty("data")
     private Data data;
     @JsonProperty("timestamp")
-    private Integer timestamp;
+    private Long timestamp;
 
     @JsonProperty("data")
     public Data getData()
@@ -31,13 +31,13 @@ public class RateResponseEntity
     }
 
     @JsonProperty("timestamp")
-    public Integer getTimestamp()
+    public Long getTimestamp()
     {
         return timestamp;
     }
 
     @JsonProperty("timestamp")
-    public void setTimestamp(Integer timestamp)
+    public void setTimestamp(Long timestamp)
     {
         this.timestamp = timestamp;
     }
